+++ src/private/release/GetLocalModulePathEntryList.ps1
@@ -1,8 +1,13 @@
 function Get-LocalModulePathEntryList {
     $separator = [IO.Path]::PathSeparator
+    $modulePath = Get-NovaEnvironmentVariableValue -Name 'PSModulePath'
+
+    if ( [string]::IsNullOrWhiteSpace($modulePath)) {
+        return @()
+    }
 
     return @(
-    $env:PSModulePath -split $separator |
+    $modulePath -split $separator |
             ForEach-Object {$_.Trim()} |
             Where-Object {-not [string]::IsNullOrWhiteSpace($_)} |
             Select-Object -Unique